hf_space/requirements.txt
torch>=2.2,<2.6
torchvision>=0.17,<0.21
gradio>=5.20,<6.0
gradio_client>=1.5.0
huggingface_hub>=0.25.0,<1.0
Pillow>=10.0.0
numpy>=1.26.0,<2.0
matplotlib>=3.8.0
